Analysis of RNAseq from BRCA data
This is the script I sent my group for the analysis of the BRCA data. I made it so that it could make scatter plots for the hypoxia associated genes. If I am going to use this in the future I am going to change it a little bit so that it doesn't make the graphs. It also needs to be relabeled so that it does not say BRCA, but I have another script for that.